requirements.txt
flask
opencv-python
pytesseract
pdf2image
pillow
face_recognition
numpy
scikit-learn
python-dotenv
